backend/trading.py
import asyncio
import logging
from typing import Dict, Any, Optional
from binance import AsyncClient
from binance.exceptions import BinanceAPIException

from config import BINANCE_API_KEY, BINANCE_SECRET_KEY, BINANCE_TESTNET

logger = logging.getLogger(__name__)

class BinanceTrading:
    def __init__(self):
        self.client: Optional[AsyncClient] = None
        self.user_positions: Dict[str, Dict] = {}
        
    async def init_client(self, api_key: str, secret_key: str):
        """바이낸스 클라이언트 초기화"""
        try:
            self.client = AsyncClient(
                api_key=api_key,
                api_secret=secret_key,
                testnet=BINANCE_TESTNET
            )
            # 연결 테스트
            await self.client.ping()
            logger.info("Binance client initialized successfully")
            return True
        except BinanceAPIException as e:
            logger.error(f"Binance API error: {e}")
            return False
        except Exception as e:
            logger.error(f"Client initialization error: {e}")
            return False
    
    async def get_account_info(self) -> Dict[str, Any]:
        """계정 정보 가져오기"""
        if not self.client:
            raise ValueError("Client not initialized")
        
        try:
            account = await self.client.futures_account()
            return account
        except Exception as e:
            logger.error(f"Failed to get account info: {e}")
            return {}
    
    async def get_positions(self) -> list:
        """현재 포지션 가져오기"""
        if not self.client:
            raise ValueError("Client not initialized")
        
        try:
            positions = await self.client.futures_position_information()
            # 비어있지 않은 포지션만 필터링
            active_positions = [pos for pos in positions if float(pos['positionAmt']) != 0]
            return active_positions
        except Exception as e:
            logger.error(f"Failed to get positions: {e}")
            return []
    
    async def place_order(self, symbol: str, side: str, quantity: float, 
                        order_type: str = "MARKET", **kwargs) -> Dict[str, Any]:
        """주문 실행"""
        if not self.client:
            raise ValueError("Client not initialized")
        
        try:
            order = await self.client.futures_create_order(
                symbol=symbol,
                side=side,
                type=order_type,
                quantity=quantity,
                **kwargs
            )
            logger.info(f"Order placed: {order}")
            return order
        except Exception as e:
            logger.error(f"Failed to place order: {e}")
            return {}
    
    async def close_position(self, symbol: str) -> bool:
        """포지션 청산"""
        try:
            positions = await self.get_positions()
            for pos in positions:
                if pos['symbol'] == symbol:
                    side = "SELL" if float(pos['positionAmt']) > 0 else "BUY"
                    quantity = abs(float(pos['positionAmt']))
                    await self.place_order(symbol, side, quantity)
                    logger.info(f"Position closed for {symbol}")
                    return True
            return False
        except Exception as e:
            logger.error(f"Failed to close position {symbol}: {e}")
            return False

# 전역 트레이딩 인스턴스
trading_instance = BinanceTrading()

async def init_binance(api_key: str, secret_key: str) -> bool:
    """바이낸스 초기화"""
    return await trading_instance.init_client(api_key, secret_key)

async def monitor_loop(user_id: str, settings: Dict[str, Any]):
    """모니터링 루프"""
    while True:
        try:
            # 여기에 시그널 분석 로직 추가
            await asyncio.sleep(10)  # 10초마다 체크
        except Exception as e:
            logger.error(f"Monitor loop error: {e}")
            await asyncio.sleep(30)  # 에러 시 30초 대기

async def position_cleanup_loop(user_id: str):
    """포지션 정리 루프"""
    while True:
        try:
            # 여기에 포지션 관리 로직 추가
            await asyncio.sleep(60)  # 1분마다 체크
        except Exception as e:
            logger.error(f"Position cleanup error: {e}")
            await asyncio.sleep(60)
