requirements.txt
fastapi
uvicorn
pandas
numpy
requests
python-dotenv
urllib3
supabase
pydantic
python-binance
aiohttp
websockets
cryptography
